requirements.txt

Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Cors


psycopg2-binary


python-dotenv
gunicorn
requests

